Change z-position for Thunderbird Park
Some buildings on the Thunderbird Park are being drawn before Thunderbird Park, and since Thunderbird Park is bigger it covers them.

So when someone presses any of the smaller buildings they get 'Thunderbird Park' details rather than the smaller building details.

I switched the order to make Thunderbird Park be first before the 4 smaller buildings that are on the same space.

Now with this change, and using certain geojson readers (leafletjs, default github geojson reader, mapbox), this order works better as it won't overlay the bigger on top of the smaller allowing the smaller buildings to be clickable.